+++ src/store/tool/selectors.ts
@@ -1,3 +1,0 @@
-export { customPluginSelectors } from './slices/customPlugin/selectors';
-export { pluginSelectors } from './slices/plugin/selectors';
-export { pluginStoreSelectors } from './slices/store/selectors';